DEX-1092-sttp3

+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/docker/DexContainer.scala
@@ -59,7 +59,7 @@ final case class DexContainer private (override val internalIp: String, underlyi
     val r = Iterator
       .continually {
         Thread.sleep(1000)
-        try httpApi.getOrderBooks.code == StatusCodes.OK
+        try httpApi.getOrderBooks.code == StatusCodes.OK.intValue
         catch {
           case _: Throwable => false
         }
